fixing torch error, removed eager enable and tf check
